requirements.txt
absl-py==1.2.0
annoy==1.17.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==5.2.0
CacheToolsUtils==6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.1.0
cloudpickle==2.1.0
contourpy==1.0.5
cryptography==37.0.4
cycler==0.11.0
Cython==0.29.32
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
Farama-Notifications==0.0.4
fasteners==0.17.3
fastjsonschema==2.16.1
filelock==3.13.1
flit_core==3.7.1
fonttools==4.34.4
fsspec==2023.12.1
future==0.18.2
glfw==2.5.4
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.47.0
gym==0.20.0
gym-notices==0.0.7
gymnasium==0.28.1
idna==3.3
imageio==2.21.0
importlib-metadata==4.11.4
importlib-resources==5.9.0
iniconfig==1.1.1
ipykernel==6.9.1
ipython==8.4.0
ipython-genutils==0.2.0
jax-jumpy==1.0.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.5
jsonschema==4.9.1
jupyter-client==7.2.2
jupyter-core==4.10.0
jupyter-server==1.18.1
jupyterlab==3.4.5
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
kiwisolver==1.4.4
llvmlite==0.39.0
lockfile==0.12.2
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
mujoco==2.3.0
mujoco-py==2.1.2.14
munkres==1.1.4
natsort==8.4.0
nbclassic==0.4.3
nbclient==0.5.13
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.5
noisyopt==0.2.2
notebook==6.4.12
notebook-shim==0.1.0
numba==0.56.0
numexpr==2.8.4
numpy==1.22.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
opencv-python==4.7.0.72
packaging==21.3
pacmap==0.6.5
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==22.2.2
pkgutil_resolve_name==1.3.10
prometheus-client==0.14.1
prompt-toolkit==3.0.20
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
PyOpenGL==3.1.6
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
Rtree==1.0.0
scikit-learn==1.1.1
scikit-optimize==0.9.0
scipy==1.9.0
seaborn==0.12.2
Send2Trash==1.8.0
setuptools==65.1.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stable-baselines3==2.2.1
stack-data==0.2.0
sympy==1.12
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
termplotlib==0.3.9
threadpoolctl==3.1.0
tinycss2==1.1.1
torch==2.1.1
torchsummary==1.5.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
trimesh==3.12.9
triton==2.1.0
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.2.1
wheel==0.37.1
zipp==3.8.1
